misc: be more pedantic about C++
* child.cc: use `nullptr` instead of `NULL`
* child_manager.cc: explicitly cast `__u32` to `pid_t`
* child_tty.hh: explicit destructor override
* event_manager.hh: move field in struct to avoid padding
* io.cc: explicitly cast `ssize_t` to `size_t`
* locker_mgr.hh: explicit destructor override
* signal_event_handler.cc: cast `__u32` -> `pid_t`
* utils.hh: use `[[maybe_unused]]` instead of `(void)`

+++ src/io.cc
@@ -41,7 +41,7 @@ size_t io::write(int fd, const void *buf, size_t count) {
 	if (r < 0)
 		throw make_system_error("write");
 
-	return r;
+	return static_cast<size_t>(r);
 }
 
 size_t io::read(int fd, void *buf, size_t count) {
@@ -53,5 +53,5 @@ size_t io::read(int fd, void *buf, size_t count) {
 	if (r < 0)
 		throw make_system_error("read");
 
-	return r;
+	return static_cast<size_t>(r);
 }
